+++ src/numerical/mod.rs
@@ -1,2 +1,4 @@
 pub mod interp;
 pub mod spline;
+pub mod newton;
+pub mod utils;